requirements.txt
discord.py==1.3.3
aiml==46.1.3